fix: weclomeページに移動
